requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-migrate==4.0.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
werkzeug==3.0.3
anthropic==0.28.0
openpyxl==3.1.4
reportlab==4.2.2
pandas==2.2.2
gunicorn==22.0.0
